Begin migrating parsing to its own file
