+++ sound_class.py
@@ -1,4 +1,5 @@
 
+from io import FileIO
 from itertools import product
 from typing import Dict, Iterable, List
 import regex as re
